# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nexo API",
        "description": "Fiscal centralizing API for integration between sales systems, government and accounting platforms.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        },
        {
            "url": "https://nexo.test",
            "description": "Herd Local Server (HTTPS)"
        },
        {
            "url": "https://backendnexo.shop",
            "description": "Mounted Server"
        }
    ],
    "paths": {
        "/api/agent/token": {
            "post": {
                "tags": [
                    "Agent"
                ],
                "summary": "Create an agent token for an organization",
                "description": "Generates a one-time-visible Bearer token that the nexo-agent daemon uses to authenticate. Store it immediately — it cannot be recovered.",
                "operationId": "agentCreateToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "organization_id"
                                ],
                                "properties": {
                                    "organization_id": {
                                        "description": "ID of the organization the agent will represent.",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "description": "Optional label to identify this token (max 100 chars).",
                                        "type": "string",
                                        "example": "nexo-agent-laptop"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Token created. Copy the token value — it will not be shown again.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Agent token created"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "description": "Plain-text token. Pass as Bearer in Authorization header. Only shown once.",
                                                    "type": "string",
                                                    "example": "a1b2c3d4e5f6..."
                                                },
                                                "agent_token_id": {
                                                    "description": "Internal ID of the created token record.",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error — organization_id missing or not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/agent/config": {
            "get": {
                "tags": [
                    "Agent"
                ],
                "summary": "Fetch Reverb connection parameters",
                "description": "Called by the daemon at startup to get the WebSocket server details. This way the agent only needs the nexo_url and its token — no infrastructure details need to be distributed manually.",
                "operationId": "agentConfig",
                "responses": {
                    "200": {
                        "description": "Reverb connection parameters.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "reverb_app_key": {
                                                    "type": "string",
                                                    "example": "drt0nv5lmajiwoj2a58h"
                                                },
                                                "reverb_host": {
                                                    "type": "string",
                                                    "example": "localhost"
                                                },
                                                "reverb_port": {
                                                    "type": "integer",
                                                    "example": 8080
                                                },
                                                "reverb_scheme": {
                                                    "type": "string",
                                                    "example": "http"
                                                },
                                                "organization_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid agent token."
                    }
                },
                "security": [
                    {
                        "agentToken": []
                    }
                ]
            }
        },
...
        "/api/agent/pending": {
            "get": {
                "tags": [
                    "Agent"
                ],
                "summary": "Reconciliation — get pending fiscal tasks",
                "description": "Called by the agent immediately after reconnecting to Reverb. Returns all sales with status=pending_fiscal for this organization, with their TaxCore-formatted payloads ready to send to V-SDC.",
                "operationId": "agentPending",
                "responses": {
                    "200": {
                        "description": "List of pending fiscalization tasks.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pending fiscal tasks"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "task_id": {
                                                        "type": "string",
                                                        "format": "uuid",
                                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                                    },
                                                    "sale_id": {
                                                        "type": "integer",
                                                        "example": 42
                                                    },
                                                    "method": {
                                                        "type": "string",
                                                        "example": "POST"
                                                    },
                                                    "endpoint": {
                                                        "type": "string",
                                                        "example": "/api/v3/invoices"
                                                    },
                                                    "payload": {
                                                        "description": "TaxCore-formatted invoice payload ready to send to V-SDC.",
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired agent token."
                    }
                },
                "security": [
                    {
                        "agentToken": []
                    }
                ]
            }
        },
...
        "/api/agent/broadcasting-auth": {
            "post": {
                "tags": [
                    "Agent"
                ],
                "summary": "Authenticate Reverb private channel",
                "description": "Called by the nexo-agent daemon after connecting to the Reverb WebSocket. Validates the agent token and returns a Pusher HMAC signature so the daemon can subscribe to private-agent.{org_id}.",
                "operationId": "agentBroadcastingAuth",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "socket_id",
                                    "channel_name"
                                ],
                                "properties": {
                                    "socket_id": {
                                        "description": "Socket ID provided by Reverb in the pusher:connection_established event.",
                                        "type": "string",
                                        "example": "123.456"
                                    },
                                    "channel_name": {
                                        "description": "Must be private-agent.{organization_id} matching the token's org.",
                                        "type": "string",
                                        "example": "private-agent.1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Auth signature generated. Use the auth value in pusher:subscribe.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "auth": {
                                                    "description": "Pusher auth string: appKey:HMAC-SHA256(socketId:channel).",
                                                    "type": "string",
                                                    "example": "nexo-local-key:a1b2c3d4..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid agent token."
                    },
                    "403": {
                        "description": "channel_name does not match this token's organization."
                    }
                },
                "security": [
                    {
                        "agentToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "agentToken": {
                "type": "http",
                "description": "Bearer token issued by POST /api/agent/token. Used by the nexo-agent daemon to authenticate all agent endpoints.",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer JWT issued by POST /api/auth/login. Pass as: Authorization: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "connectorToken": {
                "type": "http",
                "description": "Bearer token issued when a Connector is created. Only visible at creation time.",
                "scheme": "bearer"
            }
        }
    },
...
}